% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{buildYahooFinanceURL}
\alias{buildYahooFinanceURL}
\title{Build basic Yahoo Finance URL to scrape}
\usage{
buildYahooFinanceURL(ticker, urlsection)
}
\arguments{
\item{ticker}{character}

\item{urlsection}{character section name of the URL}
}
\value{
character url
}
\description{
Build basic Yahoo Finance URL to scrape
}
