% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convertUNIX2date}
\alias{convertUNIX2date}
\title{convert UNIX  to Date}
\usage{
convertUNIX2date(UNIXvalue)
}
\arguments{
\item{UNIXvalue}{}
}
\value{
Date
}
\description{
convert UNIX  to Date
}
\examples{
convertUNIX2date("1564351200")
}
