% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getYF_functions.R
\name{getYFFinancialsBasics}
\alias{getYFFinancialsBasics}
\title{Returns Yahoo Finance Basic Financials of a set of companies (tickers)
To avoid heavy loads, by default, each query gets executed every second and the system stops 10 seconds after each 10 tickers.}
\usage{
getYFFinancialsBasics(tickers, block = 10, slp = 10)
}
\arguments{
\item{block}{'integer' After each block the query stops slp seconds. Default = 10}

\item{slp}{'integer' Number os seconds to sleep after each block. Default = 10}
}
\value{
list of tickers with the Basic Financials info
}
\description{
Returns Yahoo Finance Basic Financials of a set of companies (tickers)
To avoid heavy loads, by default, each query gets executed every second and the system stops 10 seconds after each 10 tickers.
}
\examples{
getYFFinancialsBasics(c("UNA.AS", "G.MI", "BMED.MI", "wrongticker","VOW3.DE"))
}
\references{
https://finance.yahoo.com/quote/VOW3.DE/key-statistics?p=VOW3.DE
}
