% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getYF_functions.R
\name{getYFHistPrices}
\alias{getYFHistPrices}
\title{Returns Historical Prices in a data.frame
Fields: Date, volume, close, adjclose}
\usage{
getYFHistPrices(tickers, range, periodicity, block = 10, slp = 10)
}
\arguments{
\item{tickers}{set of tickers}

\item{range}{character Valid Ranges: 1y, 2y, 5y, 10y, ytd, max}

\item{periodicity}{character Valid Intervals: 1d, 1wk, 1mo}

\item{block}{}

\item{slp}{}
}
\description{
Returns Historical Prices in a data.frame
Fields: Date, volume, close, adjclose
}
\examples{
getYFHistPrices(c("UNA.AS", "G.MI", "BMED.MI", "wrongticker","VOW3.DE", "PUM.DE"), "1y", "1mo")
}
