% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getYF_functions.R
\name{getYFIndexComp}
\alias{getYFIndexComp}
\title{Get components of one index
Find list of available index calling 
Note that due to a Yahoo Finance limitation, only the top 30 are returned.
For that reason, the componenets of some indexes are taken from other sources, mainly wikipedia.}
\usage{
getYFIndexComp(index)
}
\arguments{
\item{index}{character Index ticker}
}
\value{

}
\description{
Get components of one index
Find list of available index calling 
Note that due to a Yahoo Finance limitation, only the top 30 are returned.
For that reason, the componenets of some indexes are taken from other sources, mainly wikipedia.
}
\examples{
getYFIndexComp("SP500")
}
