% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getYF_functions.R
\name{getYFProfile}
\alias{getYFProfile}
\title{Returns Yahoo Finance Profile of a set of companies (tickers)
Fields: city, country, sector, industry, website and fullTimeEmployees
To avoid heavy loads, by default, each query gets executed every second and the system stops 10 seconds after each 10 tickers.}
\usage{
getYFProfile(tickers, block = 10, slp = 10)
}
\arguments{
\item{block}{'integer' After each block the query stops slp seconds. Default = 10}

\item{slp}{'integer' Number os seconds to sleep after each block. Default = 10}
}
\value{
list of tickers with the profile info
}
\description{
Returns Yahoo Finance Profile of a set of companies (tickers)
Fields: city, country, sector, industry, website and fullTimeEmployees
To avoid heavy loads, by default, each query gets executed every second and the system stops 10 seconds after each 10 tickers.
}
\examples{
getYFProfile(c("UNA.AS", "G.MI", "BMED.MI", "wrongticker","VOW3.DE"))
}
\references{
https://finance.yahoo.com/quote/VOW3.DE/profile?p=VOW3.DE
}
