% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{queryYFchart}
\alias{queryYFchart}
\title{Query Yahoo Finance chart for one single ticker.
"chart" is the section where the historical prices are stored.
Returns a data.frame with "Date" column as character and "volume", "close" and "adjclose" as numeric.}
\usage{
queryYFchart(ticker, range, interval, from = NULL, time_in_seconds = 0.5)
}
\arguments{
\item{range}{character Valid Ranges: 1d, 5d, 1mo, 3mo, 6mo, 1y, 2y, 5y, 10y, ytd, max}

\item{interval}{character Valid Intervals: 1d, 1wk, 1mo}

\item{from}{Date "YYYY-MM-DD" from which the range counts back. Default Sys.Date}

\item{time_in_seconds}{}
}
\description{
Query Yahoo Finance chart for one single ticker.
"chart" is the section where the historical prices are stored.
Returns a data.frame with "Date" column as character and "volume", "close" and "adjclose" as numeric.
}
\examples{
queryYFchart("SAN.MC", "1y", "1wk", from = "2020-08-30")
}
\references{
:
https://github.com/dennislwy/YahooFinanceAPI
https://www.stock-data-solutions.com/kb/how-to-load-historical-prices-from-yahoo-finance-to-excel.htm
}
