% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{queryYFquoteSummary}
\alias{queryYFquoteSummary}
\title{Query Yahoo Finance quoteSummary}
\usage{
queryYFquoteSummary(ticker, module, time_in_seconds = 0.5)
}
\description{
Queries Financial information from Yahoo Finance for a single ticker and a single module.
Modules <- c(assetProfile', 'defaultKeyStatistics', 'financialData',
'industryTrend', 'indexTrend', 'sectorTrend',
'incomeStatementHistory', 'incomeStatementHistoryQuarterly',
'balanceSheetHistory', 'balanceSheetHistoryQuarterly',
'cashflowStatementHistory','cashflowStatementHistoryQuarterly',
'earnings', 'earningsHistory', 'earningsTrend',
'calendarEvents', 'secFilings', 'recommendationTrend', 'upgradeDowngradeHistory',
'institutionOwnership', 'fundOwnership', 'majorDirectHolders', 'majorHoldersBreakdown',
'insiderTransactions', 'insiderHolders', 'netSharePurchaseActivity')
}
\examples{
queryYFquoteSummary("UNA.AS", "defaultKeyStatistics")
}
