% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{queryYFquoteSummaryMany}
\alias{queryYFquoteSummaryMany}
\title{Wrapper around queryYFquoteSummary
Pre-defined the fields to extract}
\usage{
queryYFquoteSummaryMany(tickers, cols2extract, module, block = 10, slp = 10)
}
\arguments{
\item{tickers}{}

\item{cols2extract}{}

\item{module}{}

\item{block}{}

\item{slp}{}
}
\value{
list of tickers
}
\description{
Wrapper around queryYFquoteSummary
Pre-defined the fields to extract
}
