% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{scrapeWikiIndexComponents}
\alias{scrapeWikiIndexComponents}
\title{Scrape Index Components from wikipedia
Example: https://en.wikipedia.org/wiki/List_of_S%26P_500_companies}
\usage{
scrapeWikiIndexComponents(url, componentNode)
}
\arguments{
\item{url}{wikipedia url}

\item{componentNode}{character Node where the table sits}
}
\value{
data.frame
}
\description{
Scrape Index Components from wikipedia
Example: https://en.wikipedia.org/wiki/List_of_S%26P_500_companies
}
\examples{
scrapeWikiIndexComponents("https://en.wikipedia.org/wiki/List_of_S\%26P_500_companies", "//*[@id="constituents"]")
}
