% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{scrapeYahooFinance}
\alias{scrapeYahooFinance}
\title{Scrape Yahoo Finance
Returns the info of the web in different tables}
\usage{
scrapeYahooFinance(url, time_in_seconds = 1)
}
\arguments{
\item{url}{}

\item{time_in_seconds}{}
}
\description{
Scrape Yahoo Finance
Returns the info of the web in different tables
}
