% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getYF_functions.R
\name{getYFIndexComp}
\alias{getYFIndexComp}
\title{Get components of one index
Note that due to a Yahoo Finance limittion, only the top 30 are returned.}
\usage{
getYFIndexComp(ticker)
}
\arguments{
\item{ticker}{character Indey ticker}
}
\value{

}
\description{
Get components of one index
Note that due to a Yahoo Finance limittion, only the top 30 are returned.
}
\examples{
getYFIndexComp("\%5EMDAXI")
}
